refactor: centralize news types
