+++ tsconfig.app.json
@@ -23,6 +23,7 @@
 
     "baseUrl": ".",
     "paths": {
+      "@/shared/*": ["./shared/*"],
       "@/*": ["./src/*"]
     }
   },
