+++ tsconfig.json
@@ -7,6 +7,7 @@
   "compilerOptions": {
     "baseUrl": ".",
     "paths": {
+      "@/shared/*": ["./shared/*"],
       "@/*": ["./src/*"]
     },
     "noImplicitAny": false,
