+++ vite.config.ts
@@ -17,6 +17,7 @@ export default defineConfig(({ mode }) => ({
   resolve: {
     alias: {
       "@": path.resolve(__dirname, "./src"),
+      "@/shared": path.resolve(__dirname, "./shared"),
     },
   },
   build: {
